+++ src/main/java/org/apache/mesos/scheduler/MesosToSchedulerDriverAdapter.java
@@ -3,14 +3,14 @@ package org.apache.mesos.scheduler;
 import com.google.protobuf.ByteString;
 import org.apache.mesos.protobuf.Devolver;
 import org.apache.mesos.protobuf.Evolver;
+import org.apache.mesos.v1.scheduler.JNIMesos;
 import org.apache.mesos.v1.scheduler.Mesos;
 import org.apache.mesos.v1.scheduler.Protos;
+import org.apache.mesos.v1.scheduler.V0Mesos;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
+import java.util.*;
 
 /**
  * This is an adapter from the new v1 `Mesos` + `Scheduler` interface to the old  v0 `SchedulerDriver` + `Scheduler`
@@ -19,10 +19,16 @@ import java.util.List;
  * - The various `driver.xx()` calls, creates a `Call` message and then invokes `send()` on the v1
  *   `Mesos` interface.
  */
-public class MesosToSchedulerDriverAdapter
-        implements org.apache.mesos.v1.scheduler.Scheduler, org.apache.mesos.SchedulerDriver {
+public class MesosToSchedulerDriverAdapter implements
+            org.apache.mesos.v1.scheduler.Scheduler,
+            org.apache.mesos.SchedulerDriver {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MesosToSchedulerDriverAdapter.class);
+
+    // Exponential back-off between consecutive subscribe calls.
+    private static final int MULTIPLIER = 2;
+    private static final long SEED_BACKOFF_MS = 2000;
+    private static final long MAX_BACKOFF_MS = 30000;
 
     private org.apache.mesos.Scheduler wrappedScheduler;
     private org.apache.mesos.v1.Protos.FrameworkInfo frameworkInfo;
@@ -30,6 +36,8 @@ public class MesosToSchedulerDriverAdapter
     private final org.apache.mesos.v1.Protos.Credential credential;
     private boolean registered;
     private Mesos mesos;
+    private Timer subscriberTimer;
+    private long backOffMs;
 
     public MesosToSchedulerDriverAdapter(org.apache.mesos.Scheduler wrappedScheduler,
                                          org.apache.mesos.Protos.FrameworkInfo frameworkInfo,
@@ -55,22 +63,62 @@ public class MesosToSchedulerDriverAdapter
     @Override
     public void connected(Mesos mesos) {
         LOGGER.info("Connected!");
-        Protos.Call.Builder callBuilder = Protos.Call.newBuilder()
-                .setType(Protos.Call.Type.SUBSCRIBE)
-                .setSubscribe(Protos.Call.Subscribe.newBuilder()
-                        .setFrameworkInfo(frameworkInfo)
-                        .build());
+        performReliableSubscription(mesos);
+    }
 
-        if (frameworkInfo.hasId()) {
-            callBuilder.setFrameworkId(frameworkInfo.getId());
+    public class SubscriberTask extends TimerTask {
+        private Mesos mesos;
+
+        public SubscriberTask(Mesos mesos) {
+            this.mesos = mesos;
         }
 
-        mesos.send(callBuilder.build());
+        @Override
+        public void run() {
+            LOGGER.info("Sending SUBSCRIBE call");
+            Protos.Call.Builder callBuilder = Protos.Call.newBuilder()
+                    .setType(Protos.Call.Type.SUBSCRIBE)
+                    .setSubscribe(Protos.Call.Subscribe.newBuilder()
+                            .setFrameworkInfo(frameworkInfo)
+                            .build());
+            if (frameworkInfo.hasId()) {
+                callBuilder.setFrameworkId(frameworkInfo.getId());
+            }
+            mesos.send(callBuilder.build());
+            long nextBackoff = MULTIPLIER * backOffMs;
+            backOffMs = Math.min(nextBackoff > 0 ? nextBackoff : 0, MAX_BACKOFF_MS);
+            LOGGER.info("Backing off for: {}ms", backOffMs);
+            subscriberTimer.schedule(new SubscriberTask(mesos), backOffMs);
+        }
+    }
+
+    private void performReliableSubscription(Mesos mesos) {
+        // If timer is not running, initialize it.
+        if (subscriberTimer == null) {
+            initSubscriber();
+            subscriberTimer.schedule(new SubscriberTask(mesos), backOffMs);
+        }
+    }
+
+    private void initSubscriber() {
+        LOGGER.info("Initializing reliable subscriber...");
+        backOffMs = SEED_BACKOFF_MS;
+        subscriberTimer = new Timer();
+    }
+
+    private void cancelSubscriber() {
+        LOGGER.info("Cancelling subscriber...");
+        if (subscriberTimer != null) {
+            subscriberTimer.cancel();
+            subscriberTimer.purge();
+        }
+        subscriberTimer = null;
     }
 
     @Override
     public void disconnected(Mesos mesos) {
         LOGGER.info("Disconnected!");
+        cancelSubscriber();
     }
 
     @Override
@@ -80,6 +128,7 @@ public class MesosToSchedulerDriverAdapter
 
         switch (event.getType()) {
             case SUBSCRIBED: {
+                cancelSubscriber();
                 frameworkInfo = org.apache.mesos.v1.Protos.FrameworkInfo.newBuilder(frameworkInfo)
                         .setId(_event.getSubscribed().getFrameworkId())
                         .build();
@@ -119,11 +168,13 @@ public class MesosToSchedulerDriverAdapter
             case UPDATE: {
                 final org.apache.mesos.v1.Protos.TaskStatus v1Status = _event.getUpdate().getStatus();
 
+                wrappedScheduler.statusUpdate(this, event.getUpdate().getStatus());
+
+                // Send ack only when UUID is set.
                 if (v1Status.hasUuid()) {
                     final org.apache.mesos.v1.Protos.AgentID agentId = v1Status.getAgentId();
                     final org.apache.mesos.v1.Protos.TaskID taskId = v1Status.getTaskId();
 
-                    wrappedScheduler.statusUpdate(this, event.getUpdate().getStatus());
                     mesos.send(Protos.Call.newBuilder()
                             .setType(Protos.Call.Type.ACKNOWLEDGE)
                             .setFrameworkId(frameworkInfo.getId())
@@ -186,7 +237,7 @@ public class MesosToSchedulerDriverAdapter
     @Override
     public org.apache.mesos.Protos.Status start() {
         // TODO(mohit): Prevent more than 1x start() invocation.
-        String mesosApi = System.getenv("MESOS_API");
+        String mesosApi = System.getenv("MESOS_API_VERSION");
         if (mesosApi == null) {
             mesosApi = "V0";
         }
